fix: type error

+++ src/evaluation/src/module/congestion/congestion_eval.cpp
@@ -375,9 +375,9 @@ string CongestionEval::evalLUTRUDY(CongestionNets nets, CongestionRegion region,
     } else if (net_ux - net_lx < net_uy - net_ly && net_ux - net_lx != 0) {
       aspect_ratio = std::round((net_uy - net_ly) / static_cast<double>(net_ux - net_lx));
     }
-    float l_ness = 0.f;
+    double l_ness = 0.0;
     if (pin_num < 3) {
-      l_ness = 1.f;
+      l_ness = 1.0;
     } else if (pin_num <= 15) {
       std::vector<std::pair<int32_t, int32_t>> point_set;
       for (const auto& pin : net.pins) {
@@ -385,7 +385,7 @@ string CongestionEval::evalLUTRUDY(CongestionNets nets, CongestionRegion region,
       }
       l_ness = calculateLness(point_set, net_lx, net_ux, net_ly, net_uy);
     } else {
-      l_ness = 0.5f;
+      l_ness = 0.5;
     }
 
     double hor_lutrudy = 0.0;
@@ -452,18 +452,18 @@ string CongestionEval::evalLUTRUDY(CongestionNets nets, CongestionRegion region,
   return getAbsoluteFilePath(output_path);
 }
 
-float CongestionEval::calculateLness(std::vector<std::pair<int32_t, int32_t>> point_set, int32_t net_lx, int32_t net_ux, int32_t net_ly,
-                                     int32_t net_uy)
+double CongestionEval::calculateLness(std::vector<std::pair<int32_t, int32_t>> point_set, int32_t net_lx, int32_t net_ux, int32_t net_ly,
+                                      int32_t net_uy)
 {
-  int32_t bbox = (net_ux - net_lx) * (net_uy - net_ly);
+  int64_t bbox = static_cast<int64_t>(net_ux - net_lx) * static_cast<int64_t>(net_uy - net_ly);
   int32_t r1 = calcLowerLeftRP(point_set, net_lx, net_ly);
   int32_t r2 = calcLowerRightRP(point_set, net_ux, net_ly);
   int32_t r3 = calcUpperLeftRP(point_set, net_lx, net_uy);
   int32_t r4 = calcUpperRightRP(point_set, net_ux, net_uy);
   int32_t r = std::max({r1, r2, r3, r4});
-  float l_ness;
+  double l_ness;
   if (bbox != 0) {
-    l_ness = r / static_cast<float>(bbox);
+    l_ness = static_cast<double>(r) / static_cast<double>(bbox);
   } else {
     l_ness = 1.0;
   }
@@ -530,7 +530,7 @@ int32_t CongestionEval::calcUpperRightRP(std::vector<std::pair<int32_t, int32_t>
   return r;
 }
 
-double CongestionEval::getLUT(int32_t pin_num, int32_t aspect_ratio, float l_ness)
+double CongestionEval::getLUT(int32_t pin_num, int32_t aspect_ratio, double l_ness)
 {
   int ar_index;
   if (aspect_ratio == 1) {
@@ -905,9 +905,9 @@ void CongestionEval::evalNetInfo()
       aspect_ratio = std::round((net_uy - net_ly) / static_cast<double>(net_ux - net_lx));
     }
 
-    float l_ness = 0.f;
+    double l_ness = 0.0;
     if (pin_num < 3) {
-      l_ness = 1.f;
+      l_ness = 1.0;
     } else if (pin_num <= 15) {
       std::vector<std::pair<int32_t, int32_t>> point_set;
       for (const auto& pin : net.pins) {
@@ -915,12 +915,12 @@ void CongestionEval::evalNetInfo()
       }
       l_ness = calculateLness(point_set, net_lx, net_ux, net_ly, net_uy);
     } else {
-      l_ness = 0.5f;
+      l_ness = 0.5;
     }
 
     int32_t bbox_width = net_ux - net_lx;
     int32_t bbox_height = net_uy - net_ly;
-    int64_t bbox_area = bbox_width * bbox_height;
+    int64_t bbox_area = static_cast<int64_t>(bbox_width) * static_cast<int64_t>(bbox_height);
     int32_t bbox_lx = net_lx;
     int32_t bbox_ly = net_ly;
     int32_t bbox_ux = net_ux;
@@ -957,7 +957,7 @@ int CongestionEval::findAspectRatio(std::string net_name)
   throw std::runtime_error("Aspect ratio not found for net: " + net_name);
 }
 
-float CongestionEval::findLness(std::string net_name)
+double CongestionEval::findLness(std::string net_name)
 {
   auto it = _name_lness.find(net_name);
   if (it != _name_lness.end()) {
